# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "企业级在线考试系统，支持考题管理、考试配置、考生管理、成绩统计与导出。\n认证方式：Bearer Token（通过 /api/auth/login 获取，Cookie 或 Authorization Header 传递）",
        "title": "在线考试系统 API",
        "termsOfService": "http://svip.com/terms",
        "contact": {
            "name": "API 支持",
            "email": "admin@svip.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/admin/students": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "分页查询考生，支持关键词搜索（学号/姓名）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考生管理"
                ],
                "summary": "获取考生列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码 (默认 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数 (默认 10)",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "考生列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.PagedData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "管理员手动添加一名考生（学号+姓名+初始密码）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考生管理"
                ],
                "summary": "添加单个考生",
                "parameters": [
                    {
                        "description": "考生信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.StudentImportReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "考生添加成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Student"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误或学号已存在",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/admin/students/batch": {
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "一次性导入多名考生，自动跳过已存在的学号",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考生管理"
                ],
                "summary": "批量导入考生",
                "parameters": [
                    {
                        "description": "考生列表",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.StudentImportBatchReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "导入结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.BatchImportResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/admin/students/{id}": {
            "put": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "修改指定考生的姓名和/或密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考生管理"
                ],
                "summary": "更新考生信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "考生 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "{\\",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "考生信息已更新",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "删除指定考生及其相关数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考生管理"
                ],
                "summary": "删除考生",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "考生 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "考生已删除",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "无效的ID",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "删除失败",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.BatchImportResult": {
            "type": "object",
            "properties": {
                "added": {
                    "type": "integer"
                },
                "skipped": {
                    "type": "integer"
                }
            }
        },
...
        "models.PagedData": {
            "type": "object",
            "properties": {
                "list": {},
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "models.Student": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "头像",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "姓名",
                    "type": "string"
                },
                "source": {
                    "description": "来源：import（管理员导入）/ wxwork（企微注册）",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "workNo": {
                    "description": "学号/工号（唯一标识）",
                    "type": "string"
                },
                "wxOpenId": {
                    "description": "企业微信 OpenID",
                    "type": "string"
                },
                "wxUserId": {
                    "description": "企业微信 UserID（企微登录绑定）",
                    "type": "string"
                }
            }
        },
...
        "models.StudentImportBatchReq": {
            "type": "object",
            "required": [
                "students"
            ],
            "properties": {
                "students": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.StudentImportReq"
                    }
                }
            }
        },
        "models.StudentImportReq": {
            "type": "object",
            "required": [
                "name",
                "password",
                "workNo"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "workNo": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerToken": {
            "description": "输入 \"Bearer {token}\" 或通过 Cookie exam_token 传递",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}